Cleanup handwriting synthesis example
