+++ HandwritingSynthesis/download_model.sh
@@ -1,0 +1,1 @@
+wget https://github.com/t-vi/pytorch-tvmisc/releases/download/2018-03-13/graves_handwriting_generation_2018-03-13-02-45epoch_49.hd5
